Support using timer based refresh within the program
If user provide a value in the config for refresh_time_sec that is not
0 then it will refresh every x seconds processing all datasets in the
config.

+++ models/config.go
@@ -20,7 +20,7 @@ var errParseConfigFile = "Error occurred parsing the config: %s"
 type Config struct {
 	GeckoboardAPIKey string          `yaml:"geckoboard_api_key"`
 	DatabaseConfig   *DatabaseConfig `yaml:"database_config"`
-	RefreshTimeSec   int32           `yaml:"refresh_time_sec"`
+	RefreshTimeSec   uint16          `yaml:"refresh_time_sec"`
 	Datasets         []Dataset       `yaml:"datasets"`
 }
